SHM6: Class Island - change positions from size_t to int

+++ shm/island.cpp
@@ -1,5 +1,5 @@
 #include "island.hpp"
 
-Island::Island(size_t position_X, size_t position_Y)
+Island::Island(int position_X, int position_Y)
     : position_(Coordinates(position_X, position_Y)) {
 }
